requirements.txt
python-telegram-bot==21.1.1
aiohttp==3.9.5
pandas==2.2.2
python-dotenv==1.0.1
requests==2.31.0
asyncio==3.4.3
aiofiles==23.2.1
typing-extensions==4.11.0
pycoingecko==3.1.0
solana==0.34.0